demoDS3/src/main/webapp/app/entities/ds/list/ds.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpHeaders, HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { DSService } from '../service/ds.service';

import { DSComponent } from './ds.component';

describe('DS Management Component', () => {
  let comp: DSComponent;
  let fixture: ComponentFixture<DSComponent>;
  let service: DSService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [RouterTestingModule.withRoutes([{ path: 'ds', component: DSComponent }]), HttpClientTestingModule],
      declarations: [DSComponent],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            data: of({
              defaultSort: 'id,asc',
            }),
            queryParamMap: of(
              jest.requireActual('@angular/router').convertToParamMap({
                page: '1',
                size: '1',
                sort: 'id,desc',
              })
            ),
            snapshot: { queryParams: {} },
          },
        },
      ],
    })
      .overrideTemplate(DSComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(DSComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DSService);

    const headers = new HttpHeaders();
    jest.spyOn(service, 'query').mockReturnValue(
      of(
        new HttpResponse({
          body: [{ id: 123 }],
          headers,
        })
      )
    );
  });

  it('Should call load all on init', () => {
    // WHEN
    comp.ngOnInit();

    // THEN
    expect(service.query).toHaveBeenCalled();
    expect(comp.dS?.[0]).toEqual(expect.objectContaining({ id: 123 }));
  });

  describe('trackId', () => {
    it('Should forward to dSService', () => {
      const entity = { id: 123 };
      jest.spyOn(service, 'getDSIdentifier');
      const id = comp.trackId(0, entity);
      expect(service.getDSIdentifier).toHaveBeenCalledWith(entity);
      expect(id).toBe(entity.id);
    });
  });
});
